+++ static/views/cart.ejs
@@ -15,6 +15,8 @@
     <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.3.0/socket.io.dev.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
     <script defer type="text/javascript" src="../js/cart.js"></script>
+    <script type="text/javascript" src="../js/modal.js"></script>
+    <link href="../css/modal.css" rel="stylesheet" type="text/css" />
     <link rel="stylesheet" href="../css/cart.css" type="text/css">
     <link rel="stylesheet" href="../css/header.css" type="text/css">
     <link rel="stylesheet" href="../css/chatroom.css" type="text/css">
@@ -39,6 +41,7 @@
         </div>
     </section>
     <%-include('component/sidebar.ejs')%>
+    <%-include('component/modal.ejs')%>
 </body>
 <script src="../js/sticky-bar.js" type="text/javascript"></script>
 <script src="../js/sidebar.js" type="text/javascript"></script>
