+++ static/views/component/header.ejs
@@ -11,7 +11,7 @@
 				<li><a href="/manage_user">고객관리</a></li>
 				<li><a href="/users/logout">로그아웃</a></li>
 				<%} else { %>
-				<li><a href="/page/mypage">마이페이지</a></li>
+				<li><a href="/users /mypage">마이페이지</a></li>
 				<li><a href="/users/cart">장바구니</a></li>
 				<li><a href="/users/logout">로그아웃</a></li>
 				<%}%>
