+++ static/views/my_page.ejs
@@ -40,9 +40,15 @@
                 <button class="modal_button" onclick="edit_btn()">수정</button>
                 <a class="modal_close_btn">X</a>
             </div>
-            <button id="popup_open_btn">회원정보 수정</button>
+            <div class="patch_btn">
+                <button id="popup_open_btn">회원정보 수정</button>
+            </div>
         </div>
     </section>
+    <h1>구입목록</h1>
+    <h3>주문번호</h3>
+    <h3>상품명</h3>
+    <h3>수량</h3>
     <%-include('component/modal.ejs')%>
 </body>
 <script type="text/javascript" src="../js/modal.js"></script>
