+++ static/views/product_detail.ejs
@@ -17,12 +17,15 @@
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
 
     <script defer type="text/javascript" src="../js/product_detail.js"></script>
-
+    <script type="text/javascript" src="../js/modal.js"></script>
+    
+    <link href="../css/modal.css" rel="stylesheet" type="text/css" />
     <link rel="stylesheet" href="../css/home.css" type="text/css">
     <link rel="stylesheet" href="../css/header.css" type="text/css">
     <link rel="stylesheet" href="../css/chatroom.css" type="text/css">
     <link rel="stylesheet" href="../css/sidebar.css" type="text/css">
     <link rel="stylesheet" href="../css/product_detail.css" type="text/css">
+    
 
 </head>
 
@@ -36,6 +39,7 @@
         </section>
         <%-include('component/sidebar.ejs')%>
             <script src="../js/sticky-bar.js"></script>
+    <%-include('component/modal.ejs')%>
 </body>
 
 </html>
